requirements.txt
streamlit
pandas
numpy
matplotlib
datetime
python-dateutil
plotly
PythonTurtle
tk
snowflake-connector-python
jinja2
